% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_auth.R
\name{auth_google}
\alias{auth_google}
\title{Perform manual authentication with specific Google application using
self-generated specified authentication credentials}
\usage{
auth_google(appname, key, secret)
}
\arguments{
\item{appname}{Application name}

\item{key}{Google API key}

\item{secret}{Client secret}
}
\value{
A Google-authenticated session
}
\description{
Perform manual authentication with specific Google application using
self-generated specified authentication credentials
}
\examples{
\dontrun{
  auth_google(appname = "googledrive",
              key = "ADD/KEY/HERE",
              secret = "ADD/SECRET/HERE")
}

}
