% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_popup.R
\name{create_popup_content}
\alias{create_popup_content}
\title{Create Leaflet marker popup content}
\usage{
create_popup_content(.data)
}
\arguments{
\item{.data}{A data.frame containing information corresponding to each mapped
point location.}
}
\value{
A vector of character values with length equal to the number of
mapped point locations
}
\description{
Given point location data, this function facilitates the creation of
corresponding vector of character values to show in the popup for Leaflet
markers.
}
\examples{
\dontrun{
  df <- read_kb_experience()
  df <- get_kb_coordinates(.data = df)
  create_popup_content(.data = df)
}

}
