% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katilingban.R
\docType{package}
\name{katilingban}
\alias{katilingban}
\alias{katilingban-package}
\title{General Purpose Functions for Katilingban}
\description{
To support general and non-specific organisational tasks requiring or
supported by R, this package provides general purpose functions that
facilitate performant and efficient implementation of standardised workflows.
This is particularly useful for website update, newsletter generation,
reports, notes and other related tasks that are or will be automated or
supported within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/katilingban}
  \item Report bugs at \url{https://github.com/katilingban/katilingban/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Abdoul Aziz Goza
  \item Emmanuel Mandalazi
  \item Lio Fieschi
}

}
\keyword{internal}
