% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kb_experience}
\alias{kb_experience}
\title{Katilingban consultants' experience list}
\format{
A tibble with 141 rows and 12 columns
}
\usage{
kb_experience
}
\description{
Katilingban consultants' experience list
}
\examples{
kb_experience

}
\keyword{datasets}
