% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kb_sheets.R
\name{read_kb_experience}
\alias{read_kb_experience}
\title{Read Katilingban Experience Google Sheets}
\usage{
read_kb_experience(
  ss = "1l9ov8NEND-gROprl4gR5K2iiTgcQfrGirItz9uExKl0",
  sheet = 1
)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: ID, URL, etc. Set to
1l9ov8NEND-gROprl4gR5K2iiTgcQfrGirItz9uExKl0 for the Katilingban
Experience Google Sheets}

\item{sheet}{Worksheet in Google Sheets to read. Can be specified as the
name of the worksheet or by position using a number.}
}
\value{
A dribble/tibble of the Katilingban Experience Google Sheets
}
\description{
Read Katilingban Experience Google Sheets
}
\examples{
\dontrun{
  read_kb_experience()
}

}
