% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kb_sheets.R
\name{read_kb_sheets}
\alias{read_kb_sheets}
\title{Read a specified Katilingban Google Sheets}
\usage{
read_kb_sheets(ss, sheet = NULL, ...)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet: ID, URL, etc.}

\item{sheet}{Worksheet in Google Sheets to read. Can be specified as the
name of the worksheet or by position using a number.}

\item{...}{Additional parameters passed to googlesheets4::read_sheet}
}
\value{
A dribble/tibble of the specified Google Sheet
}
\description{
Read a specified Katilingban Google Sheets
}
\examples{
\dontrun{
  read_kb_sheets(ss = "1l9ov8NEND-gROprl4gR5K2iiTgcQfrGirItz9uExKl0")
}

}
